backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
dj-database-url==2.2.0
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
openai==2.36.0
google-auth==2.38.0
httpx==0.28.1
stripe==10.5.0
gunicorn==22.0.0
whitenoise==6.7.0
